backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.1
pydantic-settings==2.8.1
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
PyJWT==2.10.1
requests==2.32.3
prometheus-fastapi-instrumentator==7.1.0
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-sqlalchemy==0.52b1
pytest==8.3.5
httpx==0.28.1
ruff==0.11.0
pytest-asyncio==0.25.3
